Update backupDirectory.sh
Added functionality for flags (-d for directory and -e for email to alert out too) and cleaned up the way the script checked and created /dirbackups/.  The script now is smart enough to verify if the directory has been created and will move on with its functions.  The script now names each md5sum file with the directory it is taking the md5sum for so this script can be used for more than one directory and the has values will be matched correctly each time the script is run.